Address file-op review hardening
