+++ src/Core/bridge/apps/explorer/src/main.rs
@@ -1,8 +1,9 @@
+mod file_ops;
+
 use rayon::prelude::*;
 use std::cmp::Ordering;
 use std::env;
 use std::fs;
-use std::io::{self, Write};
 #[cfg(unix)]
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
@@ -54,7 +55,7 @@ fn run() -> Result<(), String> {
         Some("remount") => run_remount_cmd(&args[2..]),
         Some("warm-thumbnails") => run_warm_thumbnails(&args[2..]),
         Some("install-appimage") => run_install_appimage(&args[2..]),
-        Some("file-op") => run_file_op(&args[2..]),
+        Some("file-op") => file_ops::run(&args[2..]),
         _ if args.len() >= 6 => run_list(&args[1..]),
         _ => Err("usage: explorer_backend list|search|devices|mount|unmount|remount|warm-thumbnails|install-appimage|file-op ...".into()),
     }
@@ -268,320 +269,6 @@ fn run_install_appimage(args: &[String]) -> Result<(), String> {
         escape(&desktop_path.to_string_lossy())
     );
     Ok(())
-}
-
-fn run_file_op(args: &[String]) -> Result<(), String> {
-    match run_file_op_inner(args) {
-        Ok(()) => Ok(()),
-        Err(err) => {
-            println!("ERROR|{}", err);
-            flush_stdout();
-            Err(err)
-        }
-    }
-}
-
-fn run_file_op_inner(args: &[String]) -> Result<(), String> {
-    if args.len() < 5 {
-        return Err("usage: explorer_backend file-op <copy|move|cut> <destination> <overwrite|skip|rename|keep-both> <rename> <paths...>".into());
-    }
-
-    let mode = normalize_file_op_mode(&args[0])?;
-    let destination = Path::new(&args[1]);
-    let policy = args[2].as_str();
-    let rename = args[3].trim();
-    let sources: Vec<PathBuf> = args[4..].iter().map(PathBuf::from).collect();
-
-    if !destination.is_dir() {
-        return Err(format!(
-            "destination is not a directory: {}",
-            destination.display()
-        ));
-    }
-
-    println!(
-        "START|{}|{}|{}",
-        mode,
-        destination.to_string_lossy(),
-        sources.len()
-    );
-    flush_stdout();
-
-    let mut completed = 0usize;
-    for source in &sources {
-        let name = source
-            .file_name()
-            .and_then(|v| v.to_str())
-            .ok_or_else(|| format!("invalid source path: {}", source.display()))?;
-        let target_name = if policy == "rename" && sources.len() == 1 && !rename.is_empty() {
-            rename
-        } else {
-            name
-        };
-        let initial_target = destination.join(target_name);
-
-        if same_path(source, &initial_target) {
-            completed += 1;
-            emit_file_op_progress(completed, sources.len(), source);
-            continue;
-        }
-
-        let Some(target) = resolve_conflict_target(&initial_target, policy)? else {
-            completed += 1;
-            emit_file_op_progress(completed, sources.len(), source);
-            continue;
-        };
-
-        if let Some(parent) = target.parent() {
-            fs::create_dir_all(parent).map_err(|e| format!("create {}: {e}", parent.display()))?;
-        }
-
-        if mode == "move" {
-            move_path(source, &target)?;
-        } else {
-            copy_path(source, &target)?;
-        }
-
-        completed += 1;
-        emit_file_op_progress(completed, sources.len(), source);
-    }
-
-    println!(
-        "DONE|{}|{}|{}",
-        destination.to_string_lossy(),
-        completed,
-        sources.len()
-    );
-    flush_stdout();
-    Ok(())
-}
-
-fn normalize_file_op_mode(mode: &str) -> Result<&'static str, String> {
-    match mode {
-        "copy" => Ok("copy"),
-        "move" | "cut" => Ok("move"),
-        other => Err(format!("unsupported file operation mode: {other}")),
-    }
-}
-
-fn emit_file_op_progress(done: usize, total: usize, source: &Path) {
-    let percent = if total == 0 {
-        100
-    } else {
-        done.saturating_mul(100) / total
-    };
-    let name = source
-        .file_name()
-        .and_then(|v| v.to_str())
-        .map(sanitize_progress_field)
-        .unwrap_or_default();
-    println!("PROGRESS|{}|{}|{}|{}", done, total, percent, name);
-    flush_stdout();
-}
-
-fn flush_stdout() {
-    let _ = io::stdout().flush();
-}
-
-fn sanitize_progress_field(value: &str) -> String {
-    value
-        .chars()
-        .map(|c| match c {
-            '|' | '\n' | '\r' => ' ',
-            c => c,
-        })
-        .collect()
-}
-
-fn resolve_conflict_target(target: &Path, policy: &str) -> Result<Option<PathBuf>, String> {
-    if !target.exists() {
-        return Ok(Some(target.to_path_buf()));
-    }
-
-    match policy {
-        "skip" => Ok(None),
-        "overwrite" => {
-            remove_existing(target)?;
-            Ok(Some(target.to_path_buf()))
-        }
-        "rename" => Err(format!(
-            "renamed target already exists: {}",
-            target.display()
-        )),
-        "keep-both" => Ok(Some(unique_path(target))),
-        other => Err(format!("unsupported conflict policy: {other}")),
-    }
-}
-
-fn unique_path(path: &Path) -> PathBuf {
-    if !path.exists() {
-        return path.to_path_buf();
-    }
-
-    let parent = path.parent().unwrap_or_else(|| Path::new(""));
-    let stem = path
-        .file_stem()
-        .and_then(|v| v.to_str())
-        .filter(|v| !v.is_empty())
-        .or_else(|| path.file_name().and_then(|v| v.to_str()))
-        .unwrap_or("item");
-    let extension = path.extension().and_then(|v| v.to_str()).unwrap_or("");
-
-    for n in 2..10_000usize {
-        let name = if extension.is_empty() {
-            format!("{stem} {n}")
-        } else {
-            format!("{stem} {n}.{extension}")
-        };
-        let candidate = parent.join(name);
-        if !candidate.exists() {
-            return candidate;
-        }
-    }
-
-    parent.join(format!("{stem} {}", unix_millis()))
-}
-
-fn unix_millis() -> u128 {
-    std::time::SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_millis())
-        .unwrap_or(0)
-}
-
-fn same_path(a: &Path, b: &Path) -> bool {
-    if a == b {
-        return true;
-    }
-    match (fs::canonicalize(a), fs::canonicalize(b)) {
-        (Ok(a), Ok(b)) => a == b,
-        _ => false,
-    }
-}
-
-fn remove_existing(path: &Path) -> Result<(), String> {
-    if path.is_dir() && !path.is_symlink() {
-        fs::remove_dir_all(path).map_err(|e| format!("remove {}: {e}", path.display()))
-    } else {
-        fs::remove_file(path).map_err(|e| format!("remove {}: {e}", path.display()))
-    }
-}
-
-fn move_path(source: &Path, target: &Path) -> Result<(), String> {
-    match fs::rename(source, target) {
-        Ok(()) => Ok(()),
-        Err(rename_err) => {
-            if let Err(copy_err) = copy_path(source, target) {
-                let cleanup = if target.exists() {
-                    remove_existing(target)
-                        .err()
-                        .map(|err| format!("; cleanup failed ({err})"))
-                        .unwrap_or_else(|| "; partial target removed".to_string())
-                } else {
-                    String::new()
-                };
-                return Err(format!(
-                    "move {} to {}: rename failed ({rename_err}); copy fallback failed ({copy_err}){cleanup}",
-                    source.display(),
-                    target.display()
-                ));
-            }
-            remove_existing(source).map_err(|remove_err| {
-                format!(
-                    "move {} to {}: rename failed ({rename_err}); cleanup failed ({remove_err})",
-                    source.display(),
-                    target.display()
-                )
-            })
-        }
-    }
-}
-
-fn copy_path(source: &Path, target: &Path) -> Result<(), String> {
-    let meta =
-        fs::symlink_metadata(source).map_err(|e| format!("metadata {}: {e}", source.display()))?;
-    if meta.file_type().is_symlink() {
-        copy_symlink(source, target)
-    } else if meta.is_dir() {
-        copy_dir_recursive(source, target)
-    } else if meta.is_file() {
-        fs::copy(source, target)
-            .map(|_| ())
-            .map_err(|e| format!("copy {} to {}: {e}", source.display(), target.display()))
-    } else {
-        Err(format!("unsupported file type: {}", source.display()))
-    }
-}
-
-#[cfg(unix)]
-fn copy_symlink(source: &Path, target: &Path) -> Result<(), String> {
-    use std::os::unix::fs::symlink;
-    let link = fs::read_link(source).map_err(|e| format!("readlink {}: {e}", source.display()))?;
-    symlink(&link, target).map_err(|e| format!("symlink {}: {e}", target.display()))
-}
-
-#[cfg(not(unix))]
-fn copy_symlink(source: &Path, target: &Path) -> Result<(), String> {
-    let link = fs::read_link(source).map_err(|e| format!("readlink {}: {e}", source.display()))?;
-    if link.is_dir() {
-        copy_dir_recursive(&link, target)
-    } else {
-        fs::copy(&link, target)
-            .map(|_| ())
-            .map_err(|e| format!("copy {} to {}: {e}", link.display(), target.display()))
-    }
-}
-
-fn copy_dir_recursive(source: &Path, target: &Path) -> Result<(), String> {
-    if is_self_or_descendant_target(source, target) {
-        return Err(format!(
-            "refusing to copy directory into itself: {} -> {}",
-            source.display(),
-            target.display()
-        ));
-    }
-
-    fs::create_dir_all(target).map_err(|e| format!("create {}: {e}", target.display()))?;
-    let meta = fs::metadata(source).map_err(|e| format!("metadata {}: {e}", source.display()))?;
-    let _ = fs::set_permissions(target, meta.permissions());
-
-    for entry in fs::read_dir(source).map_err(|e| format!("read {}: {e}", source.display()))? {
-        let entry = entry.map_err(|e| format!("read {}: {e}", source.display()))?;
-        let child_source = entry.path();
-        let child_target = target.join(entry.file_name());
-        if child_target.exists() {
-            remove_existing(&child_target)?;
-        }
-        copy_path(&child_source, &child_target)?;
-    }
-    Ok(())
-}
-
-fn is_self_or_descendant_target(source: &Path, target: &Path) -> bool {
-    if target.starts_with(source) {
-        return true;
-    }
-
-    let Ok(source_canon) = fs::canonicalize(source) else {
-        return false;
-    };
-
-    if let Ok(target_canon) = fs::canonicalize(target) {
-        return target_canon == source_canon || target_canon.starts_with(&source_canon);
-    }
-
-    let Some(parent) = target.parent() else {
-        return false;
-    };
-    let Ok(parent_canon) = fs::canonicalize(parent) else {
-        return false;
-    };
-    let target_canon = match target.file_name() {
-        Some(name) => parent_canon.join(name),
-        None => parent_canon,
-    };
-
-    target_canon == source_canon || target_canon.starts_with(&source_canon)
 }
 
 fn parse_list_args(args: &[String]) -> Result<(&Path, bool, &str, bool, bool), String> {
